partially refactored user feedback components
