+++ projects/ui-components/lf-user-feedback/feedback-suggestion-selection/feedback-suggestion-selection.component.html
@@ -1,0 +1,6 @@
+<p class="lf-text-body lf-user-feedback-description">{{localizedStrings.YOUR_FEEDBACK_SUGGESTION_HELP_IMPROVE_PRODUCT | async}}</p>
+<p class="lf-text-body lf-user-feedback-description">{{localizedStrings.TECHNICAL_ISSUES_CONTACT_ADMIN_OR_SP | async}}</p>
+<button id="lf-user-feedback-feedback-mode-button" class="lf-button feedback-suggestion-button"
+    (click)="onClickFeedback()">{{localizedStrings.I_HAVE_FEEDBACK | async}}</button>
+<button id="lf-user-feedback-suggestion-mode-button" class="lf-button feedback-suggestion-button"
+    (click)="onClickSuggestion()">{{localizedStrings.I_HAVE_SUGGESTION | async}}</button>
