+++ projects/ui-components/lf-user-feedback/feedback-suggestion-selection/feedback-suggestion-selection.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { FeedbackSuggestionSelectionComponent } from './feedback-suggestion-selection.component';
+
+describe('FeedbackSuggestionSelectionComponent', () => {
+  let component: FeedbackSuggestionSelectionComponent;
+  let fixture: ComponentFixture<FeedbackSuggestionSelectionComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ FeedbackSuggestionSelectionComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(FeedbackSuggestionSelectionComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
