+++ projects/ui-components/lf-user-feedback/user-feedback-dialog/user-feedback-dialog.component.html
@@ -1,84 +1,44 @@
 <div mat-dialog-title class="user-feedback-title">
-  <span mat-dialog-title class="lf-popup-dialog-title">{{USER_FEEDBACK_TITLE | async}}</span>
+  <span mat-dialog-title class="lf-popup-dialog-title">{{ USER_FEEDBACK_TITLE | async }}</span>
   <button class="lf-close-button" mat-dialog-close [title]="localizedStrings.CLOSE | async">
-      <span id="pop-up-dialog-close"  class="material-icons-outlined">
-      close
-    </span>
+    <span id="pop-up-dialog-close" class="material-icons-outlined"> close </span>
   </button>
 </div>
 <div *ngIf="isFirstPane" mat-dialog-content class="lf-dialog-content">
-    <p class="lf-text-body lf-user-feedback-description">{{localizedStrings.YOUR_FEEDBACK_SUGGESTION_HELP_IMPROVE_PRODUCT | async}}</p>
-    <p class="lf-text-body lf-user-feedback-description">{{localizedStrings.TECHNICAL_ISSUES_CONTACT_ADMIN_OR_SP | async}}</p>
-  <button id="lf-user-feedback-feedback-mode-button" class="lf-button feedback-suggestion-button"
-    (click)="onClickFeedback()">{{localizedStrings.I_HAVE_FEEDBACK | async}}</button>
-  <button id="lf-user-feedback-suggestion-mode-button" class="lf-button feedback-suggestion-button"
-    (click)="onClickSuggestion()">{{localizedStrings.I_HAVE_SUGGESTION | async}}</button>
+  <lib-feedback-suggestion-selection
+    (feedbackClicked)="onClickFeedback()"
+    (suggestionClicked)="onClickSuggestion()"
+  ></lib-feedback-suggestion-selection>
 </div>
-<div *ngIf="isFeedback || isSuggestion" mat-dialog-content class="user-feedback-content-container lf-dialog-content" (drop)="dropHandler($event)"
-(dragover)="dragOverHandler($event)">
-  <div class="lf-text-body lf-user-feedback-description">
-      {{isFeedback ? (localizedStrings.FOUND_SOMETHING_LIKE_DISLIKE_LET_US_KNOW | async) : (localizedStrings.DO_YOU_HAVE_IDEA_NEW_FEATURE_IMPROVEMENT_LOOK_FORWARD_TO_HEARING | async)}}
-  </div>
-  <label for="feedback-suggestion-textbox" class="lf-text-label">{{isFeedback ? (localizedStrings.TELL_US_ABOUT_EXPERIENCE | async) :
-    (localizedStrings.TELL_US_ABOUT_IDEA | async)}}</label>
-  <textarea id="feedback-suggestion-textbox" rows="12" [maxlength]="maxFeedbackLength" [(ngModel)]="feedbackTextBox"
-      class="lf-textarea lf-text-body feedback-suggestion-textbox" [placeholder]="localizedStrings.PLEASE_DO_NOT_INCLUDE_CONFIDENTIAL_OR_PERSONAL_INFO_IN_FEEDBACK | async"></textarea>
-  <span class="lf-text-body character-count">
-    <span>{{feedbackTextBox?.length ? feedbackTextBox.length : '0'}}</span>
-    <span> / </span>
-    <span>{{maxFeedbackLength}}</span>
-  </span>
-
-  <label for="image-upload" class="lf-text-label">{{ localizedStrings.UPLOAD_IMAGE_OPTIONAL | async }}</label>
-  <div id="image-upload">
-    <div *ngIf="!imageUploaded || !isImageValid" id="drop_zone" class="drop-zone">
-      <label for="upload-button" class="lf-text-body">{{ localizedStrings.DRAG_SELECT_IMAGE | async }}</label>
-      <button
-        id="upload-button"
-        [disabled]="isImageValid"
-        class="lf-button sec-button feedback-file-upload-button"
-        (click)="onInputClickArea()"
-      >
-        {{ localizedStrings.BROWSE | async }}
-      </button>
-    </div>
-
-    <tr *ngIf="isImageValid" class="upload-file-info">
-      <td>{{ imageUploaded?.name }}</td>
-      <td>{{ uploadedImageSize }}</td>
-      <td class="upload-file-remove" (click)="removeImage()">{{ localizedStrings.REMOVE | async }}</td>
-    </tr>
-
-    <input
-      type="file"
-      accept=".jpg,.png,.gif,.webp"
-      class="file-input feedback-file-input"
-      (input)="onFileSelectedAsync($event)"
-      #uploadFile
-    />
-    <p *ngIf="!isImageValid" class="error-message lf-text-body">{{ imageUploadErrorMessage | async }}</p>
-  </div>
-
-  <div class="email-checkbox-container">
-      <mat-checkbox color="primary" class="lf-user-feedback-checkbox lf-checkbox"
-          [disableRipple]="true" ngDefaultControl [(ngModel)]="feedbackEmailCheckbox">
-      {{localizedStrings.YOU_MAY_CONTACT_ME_ABOUT_FEEDBACK | async}}
-    </mat-checkbox>
-  </div>
+<div *ngIf="isFeedback || isSuggestion"
+mat-dialog-content>
+<lib-feedback-submission [isFeedback]="isFeedback"></lib-feedback-submission>
 </div>
 <div *ngIf="isThankYou" mat-dialog-content class="user-feedback-content-container lf-dialog-content">
-    <p class="lf-text-body lf-user-feedback-thank-you">{{localizedStrings.THANK_YOU_FOR_SUBMISSION | async}}{{localizedStrings.IF_YOUD_LIKE_TO_JOIN_OUR_CUSTOMER_PANEL | async}}
-      <a href="https://info.laserfiche.com/laserfiche-customer-panel-signup" target="_blank">{{localizedStrings.PLEASE_CLICK_HERE | async}}</a>
+  <p class="lf-text-body lf-user-feedback-thank-you">
+    {{ localizedStrings.THANK_YOU_FOR_SUBMISSION | async
+    }}{{ localizedStrings.IF_YOUD_LIKE_TO_JOIN_OUR_CUSTOMER_PANEL | async }}
+    <a href="https://info.laserfiche.com/laserfiche-customer-panel-signup" target="_blank">{{
+      localizedStrings.PLEASE_CLICK_HERE | async
+    }}</a>
   </p>
 </div>
 <div *ngIf="isError" mat-dialog-content class="user-feedback-content-container lf-dialog-content">
-    <p class="lf-text-body lf-user-feedback-thank-you">{{localizedStrings.SOMETHING_WENT_WRONG_PLEASE_TRY_AGAIN_LATER | async}}</p>
+  <p class="lf-text-body lf-user-feedback-thank-you">
+    {{ localizedStrings.SOMETHING_WENT_WRONG_PLEASE_TRY_AGAIN_LATER | async }}
+  </p>
 </div>
 <mat-dialog-actions align="end" class="lf-dialog-footer">
-  <button *ngIf="isFeedback || isSuggestion" id="lf-user-feedback-submit-feedback-button"
-    class="lf-button primary-button dialog-button lf-user-feedback-submit-feedback-button" [disabled]="isSubmitDisabled"
-    (click)="onClickSubmitAsync()">{{localizedStrings.SUBMIT | async}}</button>
+  <button
+    *ngIf="isFeedback || isSuggestion"
+    id="lf-user-feedback-submit-feedback-button"
+    class="lf-button primary-button dialog-button lf-user-feedback-submit-feedback-button"
+    [disabled]="isSubmitDisabled"
+    (click)="onClickSubmitAsync()"
+  >
+    {{ localizedStrings.SUBMIT | async }}
+  </button>
   <button class="lf-button sec-button dialog-button lf-user-feedback-close-feedback-button" mat-dialog-close>
-    {{(isThankYou || isError) ? (localizedStrings.CLOSE | async) : (localizedStrings.CANCEL | async)}}
+    {{ isThankYou || isError ? (localizedStrings.CLOSE | async) : (localizedStrings.CANCEL | async) }}
   </button>
 </mat-dialog-actions>
