+++ projects/ui-components/lf-user-feedback/user-feedback-dialog/user-feedback-dialog.component.ts
@@ -32,18 +32,11 @@ export enum FeedbackDialogState {
 })
 export class UserFeedbackDialogComponent implements AfterViewInit {
   @Output() submitFeedback: EventEmitter<UserFeedbackDialogData> = new EventEmitter();
-  @ViewChild('uploadFile') inputFile?: ElementRef<HTMLInputElement>;
 
   dialogState: FeedbackDialogState = FeedbackDialogState.FIRST_PANE;
   readonly maxFeedbackLength: number = 2048;
   feedbackTextBox: string = '';
   feedbackEmailCheckbox: boolean = false;
-  feedbackImageBase64: string | undefined;
-  imageUploaded?: File;
-  uploadedImageSize: string | undefined;
-  imageSizeLimitBytes: number = 2.9 * 1024 * 1024; // limit is 2.9MB
-  isImageValid: boolean = false;
-  imageUploadErrorMessage?: Observable<string>;
 
   get isSubmitDisabled(): boolean {
     return (
@@ -71,12 +64,6 @@ export class UserFeedbackDialogComponent implements AfterViewInit {
     SUGGESTION: this.localizationService.getStringLaserficheObservable('SUGGESTION'),
     FEEDBACK: this.localizationService.getStringLaserficheObservable('FEEDBACK'),
     CLOSE: this.localizationService.getStringLaserficheObservable('CLOSE'),
-    YOUR_FEEDBACK_SUGGESTION_HELP_IMPROVE_PRODUCT: this.localizationService.getStringComponentsObservable(
-      'YOUR_FEEDBACK_SUGGESTION_HELP_IMPROVE_PRODUCT'
-    ),
-    TECHNICAL_ISSUES_CONTACT_ADMIN_OR_SP: this.localizationService.getStringComponentsObservable(
-      'TECHNICAL_ISSUES_CONTACT_ADMIN_OR_SP'
-    ),
     FOUND_SOMETHING_LIKE_DISLIKE_LET_US_KNOW: this.localizationService.getStringComponentsObservable(
       'FOUND_SOMETHING_LIKE_DISLIKE_LET_US_KNOW'
     ),
@@ -90,8 +77,6 @@ export class UserFeedbackDialogComponent implements AfterViewInit {
       ),
     TELL_US_ABOUT_EXPERIENCE: this.localizationService.getStringComponentsObservable('TELL_US_ABOUT_EXPERIENCE'),
     TELL_US_ABOUT_IDEA: this.localizationService.getStringComponentsObservable('TELL_US_ABOUT_IDEA'),
-    I_HAVE_FEEDBACK: this.localizationService.getStringComponentsObservable('I_HAVE_FEEDBACK'),
-    I_HAVE_SUGGESTION: this.localizationService.getStringComponentsObservable('I_HAVE_SUGGESTION'),
     THANK_YOU_FOR_SUBMISSION: this.localizationService.getStringComponentsObservable('THANK_YOU_FOR_SUBMISSION'),
     YOU_MAY_CONTACT_ME_ABOUT_FEEDBACK: this.localizationService.getStringComponentsObservable(
       'YOU_MAY_CONTACT_ME_ABOUT_FEEDBACK'
@@ -106,10 +91,6 @@ export class UserFeedbackDialogComponent implements AfterViewInit {
     SUBMIT: this.localizationService.getStringLaserficheObservable('SUBMIT'),
     CANCEL: this.localizationService.getStringLaserficheObservable('CANCEL'),
     REMOVE: this.localizationService.getStringLaserficheObservable('REMOVE'),
-    FILE_TOO_LARGE: this.localizationService.getStringComponentsObservable('FILE_TOO_LARGE'),
-    IMAGE_CORRUPTED_FORMAT_UNRECOGNIZED: this.localizationService.getStringComponentsObservable(
-      'IMAGE_CORRUPTED_FORMAT_UNRECOGNIZED'
-    ),
     BROWSE: this.localizationService.getStringLaserficheObservable('BROWSE'),
     DRAG_SELECT_IMAGE: this.localizationService.getStringComponentsObservable('DRAG_SELECT_IMAGE'),
     UPLOAD_IMAGE_OPTIONAL: this.localizationService.getStringComponentsObservable('UPLOAD_IMAGE_OPTIONAL'),
@@ -154,82 +135,7 @@ export class UserFeedbackDialogComponent implements AfterViewInit {
     }
   }
 
-  onInputClickArea(): void {
-    if (!this.inputFile) {
-      return;
-    }
-    this.inputFile.nativeElement.click();
-  }
 
-  async onFileSelectedAsync(event: Event): Promise<void> {
-
-    // console.log(event);
-    this.imageUploaded = this.inputFile?.nativeElement.files?.item(0) ?? undefined;
-    await this.readAndValidateImageAsync();
-  }
-
-  async getBase64Async(file: File): Promise<string> {
-    return new Promise((resolve, reject) => {
-      var reader = new FileReader();
-      reader.onload = () => {
-        var image = document.createElement('img');
-        image.onload = () => {
-          const imgBase64: string = reader.result as string;
-          resolve(imgBase64);
-        };
-        image.onerror = (error) => {
-          console.error('invalid image: ', error);
-          reject(new ImageUploadError((error as string) ?? 'error event', ImageUploadErrorType.UnsupportedFormat));
-        };
-        image.src = reader.result as string;
-        //TODO: review the above cast
-        // this.feedbackImageBase64 = reader.result as string;
-      };
-      reader.onerror = (error: any) => {
-        //TODO: is message the expected property?
-        reject(new ImageUploadError(error?.message ?? 'error event', ImageUploadErrorType.UnsupportedFormat));
-        console.log('Error: ', error);
-      };
-      reader.readAsDataURL(file);
-    });
-  }
-
-  removeImage(): void {
-    // this is to clear the selection of the input element
-    this.isImageValid = false;
-    if (!this.inputFile?.nativeElement.files) {
-      return;
-    }
-    this.inputFile.nativeElement.files = null;
-    this.imageUploaded = undefined;
-    this.inputFile.nativeElement.value = '';
-  }
-
-  async dropHandler(ev: DragEvent): Promise<void> {
-    let file: File | undefined;
-    console.log('File(s) dropped');
-
-    // Prevent default behavior (Prevent file from being opened)
-    ev.preventDefault();
-
-    if (ev?.dataTransfer?.items) {
-      // Use DataTransferItemList interface to access the file
-      const item = ev.dataTransfer.items[0];
-      if (item.kind === 'file') {
-        file = item.getAsFile() ?? undefined;
-      }
-    } else {
-      // Use DataTransfer interface to access the file(s)
-      file = ev.dataTransfer?.files.item(0) ?? undefined;
-    }
-    this.imageUploaded = file;
-    await this.readAndValidateImageAsync();
-  }
-
-  dragOverHandler(ev: DragEvent) {
-    // Prevent default behavior (Prevent file from being opened)
-    ev.preventDefault();
-  }
 
   private isEmptyOrWhitespace(value: string): boolean {
     return !(value.trim().length > 0);
@@ -248,69 +154,5 @@ export class UserFeedbackDialogComponent implements AfterViewInit {
     };
     return dialogData;
   }
-  private formatBytes(bytes: number, decimals: number = 2): string {
-    if (!+bytes) return '0 Bytes';
-
-    const k = 1024;
-    const dm = decimals < 0 ? 0 : decimals;
-    const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-
-    const i = Math.floor(Math.log(bytes) / Math.log(k));
-
-    return `${parseFloat((bytes / Math.pow(k, i)).toFixed(dm))} ${sizes[i]}`;
-  }
-
-  private async readAndValidateImageAsync(): Promise<void> {
-    if (!this.imageUploaded) {
-      return;
-    }
-    this.uploadedImageSize = this.formatBytes(this.imageUploaded.size);
-    try {
-      if (this.imageUploaded.size <= this.imageSizeLimitBytes) {
-        const encodingData = await this.getBase64Async(this.imageUploaded);
-        this.feedbackImageBase64 = encodingData?.split(',')[1];
-        // console.log(this.feedbackImageBase64); // TODO: remove
-        this.isImageValid = true;
-        this.imageUploadErrorMessage = undefined;
-      } else {
-        // TODO: disable submit or notify users that the image will not be submitted
-        throw new ImageUploadError('ImageUploadErrorType.TooLarge', ImageUploadErrorType.TooLarge);
-      }
-    } catch (error: any) {
-      let errorMessage: Observable<string>;
-      if (error.name === ImageUploadError_name) {
-        switch ((<ImageUploadError>error).imageUploadErrorType)
-        {
-          case ImageUploadErrorType.TooLarge:
-            errorMessage = this.localizedStrings.FILE_TOO_LARGE;
-            break;
-          case ImageUploadErrorType.UnsupportedFormat:
-            errorMessage = this.localizedStrings.IMAGE_CORRUPTED_FORMAT_UNRECOGNIZED;
-            break;
-          default:
-            errorMessage = new Observable<string>(error.message);
-            break;
-        }
-      } else {
-        errorMessage = new Observable<string>(error.message);
-      }
-      console.log(error); // TODO: remove
-      this.isImageValid = false;
-      this.imageUploadErrorMessage = errorMessage;
-    }
-  }
 }
 
-enum ImageUploadErrorType {
-  'TooLarge',
-  'UnsupportedFormat',
-}
-
-const ImageUploadError_name = 'ImageUploadError';
-class ImageUploadError extends Error {
-  name = ImageUploadError_name;
-  constructor(message: string, public imageUploadErrorType: ImageUploadErrorType) {
-    super(message);
-    // Set the prototype explicitly.
-  }
-}
